% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_dynamic.R
\name{gen_dynamic}
\alias{gen_dynamic}
\title{Generate dynamic networks and simulate games}
\usage{
gen_dynamic(N, L, aver_k = 6, net_type = "BA", game_type = "PDG")
}
\arguments{
\item{N}{Number of nodes in the network}

\item{L}{Number of simulation steps}

\item{aver_k}{Average degree of nodes (default is 6)}

\item{net_type}{Type of network to generate ("BA", "ER", "SW")}

\item{game_type}{Type of game to simulate ("PDG", "SDG", "DG", "SUG", etc.)}
}
\value{
A list containing simulation results including accumulated payoff matrix,
initial and final adjacency matrices, and payoff matrices for each node.
}
\description{
This function generates dynamic networks based on the specified type
and simulates various types of games on the network.
}
